Pin pint to version used in development
Apparently there is a breaking change in one of the new versions for the
way that I used .units.format_babel. Passing region="EN" works in most
places, but it seems to throw errors when checking pressure units, so
pinning to dev version seemed like the way to go instead.